Unite order preserving casts for core sort and KT SYCL sort
