+++ include/oneapi/dpl/pstl/hetero/dpcpp/parallel_backend_sycl.h
@@ -1430,7 +1430,7 @@ struct __is_radix_sort_usable_for_type
 {
     static constexpr bool value =
 #if _ONEDPL_USE_RADIX_SORT
-        (::std::is_integral_v<_T> || __is_radix_sort_float_v<_T>) &&
+        (::std::is_integral_v<_T> || oneapi::dpl::__internal::__is_radix_sort_float_v<_T>) &&
         (__internal::__is_comp_ascending<::std::decay_t<_Compare>>::value ||
          __internal::__is_comp_descending<::std::decay_t<_Compare>>::value);
 #else
